feat(#40): organized files
