+++ back-end/src/model/database/database.providers.ts
@@ -9,7 +9,7 @@ export const databaseProviders = [
         host: 'localhost',
         port: 3306,
         username: 'root',
-        password: '4miL4nurB#',
+        password: '781781',
         database: 'meioameio',
         entities: [__dirname + '/../**/*.entity{.ts,.js}'],
         synchronize: true, //toda vez que rodar a aplicação, o nest vai tentar sincronizar o db com as classes
