+++ back-end/src/professor/professor.controller.spec.ts
@@ -1,0 +1,21 @@
+import { ProfessorService } from './professor.service';
+import { professorProviders } from './professor.providers';
+import { ProfessorController } from './professor.controller';
+import { Test, TestingModule } from '@nestjs/testing';  
+
+describe('ProfessorController', () => {
+    let controller: ProfessorController;
+  
+    beforeEach(async () => {
+      const module: TestingModule = await Test.createTestingModule({
+        controllers: [ProfessorController],
+        providers: [ProfessorService],
+      }).compile();
+  
+      controller = module.get<ProfessorController>(ProfessorController);
+    });
+
+    it('should be defined', () => {
+        expect(controller).toBeDefined();
+    });
+});
