+++ back-end/src/professor/professor.module.ts
@@ -1,5 +1,5 @@
 import { Module } from '@nestjs/common';
-import { DatabaseModule } from '../database/database.module';
+import { DatabaseModule } from '../model/database/database.module';
 import { ProfessorController } from './professor.controller';
 import { ProfessorService } from './professor.service';
 import { professorProviders } from './professor.providers';
